+++ testing/testing_zlaset_band.cpp
@@ -51,7 +51,7 @@ int main( int argc, char** argv)
 
     magma_uplo_t uplo[] = { MagmaLower, MagmaUpper, MagmaFull };
     
-    printf("K = nb = %d\n", nb );
+    printf("K = nb = %d\n", (int) nb );
     printf("uplo       M     N   CPU GByte/s (ms)    GPU GByte/s (ms)    check\n");
     printf("==================================================================\n");
     for( int iuplo = 0; iuplo < 2; ++iuplo ) {
